step2 리팩터링 : ReservationRepository 파일 추가하여 레포지터리 따로 관리하도록 분리
