+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,54 +1,9 @@
 package roomescape.domain;
 
-public class Reservation {
-    private Long id;
-    private String name;
-    private String date;
-    private String time;
-
-    public Reservation(Long id, String name, String date, String time) {
-        this.id = id;
-        this.name = name;
-        this.date = date;
-        this.time = time;
-    }
-
-    public boolean validate() {
-        if(this.getDate() == "" || this.getTime() == ""){
-            return false;
-        }
-        return true;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public void setDate(String date) {
-        this.date = date;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-    public void setTime(String time) {
-        this.time = time;
-    }
+public record Reservation(
+        Long id,
+        String name,
+        String date,
+        String time
+        ){
 }
